+++ CotEditor/Sources/SettingFileManaging.swift
@@ -8,7 +8,7 @@
 //
 //  ---------------------------------------------------------------------------
 //
-//  © 2016-2018 1024jp
+//  © 2016-2019 1024jp
 //
 //  Licensed under the Apache License, Version 2.0 (the "License");
 //  you may not use this file except in compliance with the License.
@@ -311,6 +311,9 @@ extension SettingFileManaging {
         { (newReadingURL, newWritingURL) in
             
             do {
+                if newWritingURL.isReachable {
+                    try FileManager.default.removeItem(at: newWritingURL)
+                }
                 try FileManager.default.copyItem(at: newReadingURL, to: newWritingURL)
                 
                 var newWritingURL = newWritingURL